removed unneeded excelController added upload/download to subscriptionController added flash messages to app.blade.php
