+++ composer.json
@@ -7,7 +7,8 @@
 	"require": {
 		"laravel/framework": "5.0.*",
 		"illuminate/html": "~5.0",
-        "maatwebsite/excel": "~2.0"
+        "maatwebsite/excel": "~2.0",
+		"laracasts/flash": "~1.3"
     },
   "require-dev": {
       "phpunit/phpunit": "~4.0",
