+++ config/app.php
@@ -136,7 +136,8 @@ return [
 		'Illuminate\Validation\ValidationServiceProvider',
 		'Illuminate\View\ViewServiceProvider',
         'Illuminate\Html\HtmlServiceProvider',
-        'Maatwebsite\Excel\ExcelServiceProvider',
+
+
 
 		/*
 		 * Application Service Providers...
@@ -146,7 +147,8 @@ return [
 		'App\Providers\ConfigServiceProvider',
 		'App\Providers\EventServiceProvider',
 		'App\Providers\RouteServiceProvider',
-
+        'Laracasts\Flash\FlashServiceProvider',
+        'Maatwebsite\Excel\ExcelServiceProvider',
 	],
 
 	/*
@@ -197,6 +199,7 @@ return [
         'Form'      => 'Illuminate\Html\FormFacade',
         'Html'      => 'Illuminate\Html\HtmlFacade',
         'Excel'     => 'Maatwebsite\Excel\Facades\Excel',
+        'Flash'     => 'Laracasts\Flash\Flash',
 
 	],
