--- resources/views/excel/download.blade.php
@@ -1,15 +1,0 @@
-@extends('app')
-
-@section ('menu')
-    @include('partials.appMenu', ['activeMenu'  => 'excel'])
-@stop()
-
-@section('content')
-
-    <div class="container">
-        <p>asfasdfasdf</p>
-    </div>
-
-
-
-@stop()
