+++ resources/views/partials/providerForm.blade.php
@@ -23,5 +23,5 @@
     {!! Form::input('date', 'valid_until', date('Y-m-d'), ['class' => 'form-control']) !!}
 </div>
 <div class="form-group">
-    {!! Form::submit($submitButtonText, ['class' => 'btn btn-primary form-control']) !!}
+    {!! Form::submit($submitButtonText, ['class' => 'btn btn-primary']) !!}
 </div>
