+++ resources/views/partials/subscriptionForm.blade.php
@@ -56,8 +56,8 @@
 </div>
 <div class="form-group">
     {!! Form::label('valid_until', 'Actief t/m: ') !!}
-    {!! Form::input('date', 'valid_until', $subscription->valid_until, ['class' => 'form-control']) !!}
+    {!! Form::input('text', 'valid_until', null, ['class' => 'form-control']) !!}
 </div>
 <div class="form-group">
-    {!! Form::submit($submitButtonText, ['class' => 'btn btn-primary form-control']) !!}
+    {!! Form::submit($submitButtonText, ['class' => 'btn btn-primary']) !!}
 </div>
