+++ resources/views/subscription/partials/form.blade.php
@@ -1,5 +1,5 @@
 <div class="form-group">
-    {!! Form::label('excel', 'Upload .xls: ') !!}
+    {!! Form::label('excel', 'Kies bestand: ') !!}
     {!! Form::file('excel', null, ['class' => 'form-control']) !!}
 </div>
 <div class="form-group">
