+++ resources/views/subscription/show.blade.php
@@ -6,10 +6,8 @@
 
 @section('content')
 
-    <div class="container">
         <h2>Provider : {{ $providerName }}</h2>
         <p>Alle {{ $providerName }} abonnementartikelen.</p>
         @include('partials.subscriptionTable')
-    </div>
 
 @stop()
